base/pyproject.toml
[project]
name = "base"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "pymongo>=4.6.1",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "authlib>=1.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "itsdangerous>=2.1.2",
    "pika>=1.3.2",
    "sqlalchemy>=2.0.24",
    "psycopg2>=2.9.9",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload"